Remove old GetBases() and dependent methods
