+++ Assets/Scripts/AI/ProofBasis.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 78f869c00c80044baa04c19fdaac7b53
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
